Changes:  -previously, if a non-number was entered in a function, the program 'misheard' the user and suggested a random number. I have removed this, despite being funny, and now have it clear the CIN and ask again infinitely.  -added CIN fail checking to the main menu and list of options.  -fixed the factorial function to only work on whole numbers.  -changed all vars & functions to use number types with the largest possible range.  -fixed the exponent function to properly calculate negative exponents.

+++ learn/functions.h
@@ -3,12 +3,12 @@
 
 #include "stdafx.h"
 
-float addNum(float x, float y);
-float subtractNum(float x, float y);
-float multiplyNum(float x, float y);
-float divideNum(float x, float y);
-float exponentNum(float x, int y);
-float factorialNum(float x);
+long double addNum(long double x, long double y);
+long double subtractNum(long double x, long double y);
+long double multiplyNum(long double x, long double y);
+long double divideNum(long double x, long double y);
+long double exponentNum(long double x, int y);
+unsigned long long factorialNum(unsigned long long x);
 unsigned long long lowBound(int exp);
 unsigned long long highBound(int exp);
